add OSValidator for zip

+++ src/Tools/archive.java
@@ -1,4 +1,4 @@
-package Tools;
+package tools;
 
 import java.io.*;
 import java.util.logging.Level;
@@ -10,11 +10,17 @@ import java.util.zip.ZipInputStream;
  * Created by John on 06.07.2015.
  */
 
-public class archive {
+public class Archive {
 
     public static void unZip(final String zipFileName) {
 
-        String destinationDirectory = "C:/Users/John/AppData/Roaming/.minecraft/";
+        String destinationDirectory = "";
+        if(OSValidator.checkOS()==1) {
+            destinationDirectory = System.getenv("APPDATA") + "/.minecraft/";
+        }
+        if(OSValidator.checkOS()==3) {
+            destinationDirectory = System.getProperty("user.home") + "/.minecraft/"; //for unix
+        }
         int BUFFER_SIZE = 1024;
 
         byte[] buffer = new byte[BUFFER_SIZE];
@@ -25,46 +31,47 @@ public class archive {
         if (!dstDir.exists()) {
             dstDir.mkdir();
         }
-
-        try {
-            // Получаем содержимое ZIP архива
-            final ZipInputStream zis = new ZipInputStream(
-                    new FileInputStream(zipFileName));
-            ZipEntry ze = zis.getNextEntry();
-            String nextFileName;
-            while (ze != null) {
-                nextFileName = ze.getName();
-                File nextFile = new File(dstDirectory + File.separator
-                        + nextFileName);
-                System.out.println("unzip file: "
-                        + nextFile.getAbsolutePath());
-                // Если мы имеем дело с каталогом - надо его создать. Если
-                // этого не сделать, то не будут созданы пустые каталоги
-                // архива
-                if (ze.isDirectory()) {
-                    nextFile.mkdir();
-                } else {
-                    // Создаем все родительские каталоги
-                    new File(nextFile.getParent()).mkdirs();
-                    // Записываем содержимое файла
-                    try (FileOutputStream fos
-                                 = new FileOutputStream(nextFile)) {
-                        int length;
-                        while ((length = zis.read(buffer)) > 0) {
-                            fos.write(buffer, 0, length);
+        if (new File(zipFileName).exists()) {
+            try {
+                // Получаем содержимое ZIP архива
+                final ZipInputStream zis = new ZipInputStream(
+                        new FileInputStream(zipFileName));
+                ZipEntry ze = zis.getNextEntry();
+                String nextFileName;
+                while (ze != null) {
+                    nextFileName = ze.getName();
+                    File nextFile = new File(dstDirectory + File.separator
+                            + nextFileName);
+                    System.out.println("unzip file: "
+                            + nextFile.getAbsolutePath());
+                    // Если мы имеем дело с каталогом - надо его создать. Если
+                    // этого не сделать, то не будут созданы пустые каталоги
+                    // архива
+                    if (ze.isDirectory()) {
+                        nextFile.mkdir();
+                    } else {
+                        // Создаем все родительские каталоги
+                        new File(nextFile.getParent()).mkdirs();
+                        // Записываем содержимое файла
+                        try (FileOutputStream fos
+                                     = new FileOutputStream(nextFile)) {
+                            int length;
+                            while ((length = zis.read(buffer)) > 0) {
+                                fos.write(buffer, 0, length);
+                            }
                         }
                     }
+                    ze = zis.getNextEntry();
                 }
-                ze = zis.getNextEntry();
+                zis.closeEntry();
+                zis.close();
+            } catch (FileNotFoundException ex) {
+                Logger.getLogger(Archive.class.getName())
+                        .log(Level.SEVERE, null, ex);
+            } catch (IOException ex) {
+                Logger.getLogger(Archive.class.getName())
+                        .log(Level.SEVERE, null, ex);
             }
-            zis.closeEntry();
-            zis.close();
-        } catch (FileNotFoundException ex) {
-            Logger.getLogger(archive.class.getName())
-                    .log(Level.SEVERE, null, ex);
-        } catch (IOException ex) {
-            Logger.getLogger(archive.class.getName())
-                    .log(Level.SEVERE, null, ex);
         }
     }
 }
